requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
tensorboard>=2.13.0
optuna>=3.0.0
matplotlib>=3.7.0
